Modify SuccessfulSwapDto types from float to BigDecimal

+++ src/main/java/com/market/tradingbit/dtos/SuccessfulSwapDto.java
@@ -1,14 +1,22 @@
 package com.market.tradingbit.dtos;
 
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
+import jakarta.persistence.Column;
 import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 @Builder
 public class SuccessfulSwapDto {
     private String from;
     private String to;
-    private float fromQuantity;
-    private float toQuantity;
-    private float fee;
+
+    private BigDecimal fromQuantity;
+
+    private BigDecimal toQuantity;
+
+    private BigDecimal fee;
 }
